// src/generated/api/models/GetAchievementByIdDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type GetAchievementByIdDto = {
    /**
     * Player UUID
     */
    uuid: string;
    /**
     * Achievement ID
     */
    achievementId: string;
};